+++ breakout-backend/src/java/nu/smashit/core/power/PowerFactory.java
@@ -14,17 +14,17 @@ public class PowerFactory {
                 && brickType.getType() != BrickType.Type.D) {
             return new NoPower();
         } else {
-            int value = powerData.getValue();
+            double value = powerData.getValue();
             int powerID = powerData.getPowerID();
             switch (powerData.getType()) {
                 case "POINTS":
                     return new Points(powerID, value);
                 case "BALL_DIAMETER":
-                    return new BallDiameter(powerID, value);
+                    return new BallDiameter(powerID, (int) value);
                 case "PALET_SIZE":
-                    return new PaletSize(powerID, value);
+                    return new PaletSize(powerID, (int) value);
                 case "START_SPEED":
-                    return new StartSpeed(powerID, value);
+                    return new StartSpeed(powerID, (int) value);
                 default:
                     return new NoPower();
             }
